Camera fix
The camera is now fixed but it's probably not an ideal fix. All I do is
set the ship's yaw to 0, assign the camera up to the ship's forward and
then apply the yaw again.

A possible better fix would be a camera class with an Up, Look at and
Right vector and recalculate the vectors as it moves.