Bullets Working
Bullets should now be working correctly. Had to apply a similar "trick"
to what I did to fix the camera. I set the bullet to have the rotation
of the player, but before doing that, I must set the yaw to 0. I then
reset the yaw for the player and use yaw from the shot angle for the
bullet.