That smoke!
Tweaked the scale of stuff somewhat as I now realise it was far too
scaled up. Also been looking into particles, was trying out Mercurey
Particle Engine and then found out it was only 2D although it has a nice
editor and stuff it looked really good.

Anyhow, I found DPSF which I am now using. It doesn't have an editor
which is a bit annoying as I'll have to tweak things and build all the
time but it has some nice examples which I should be able to adapt to
suite my needs.