Big update, read description.
So I haven't done a commit for a while. However I have been working on
it a little.

The biggest change is that the Entity structure has changed and the way
they are drawn. Each entity is now drawing itself when it is called by
the Draw class. This was due to some things like the Skybox and Planet
to be drawn differently as well as the Planet and Skybox should not have
collision.

On collision, there is now a check in Game1 checking for ocllision
between bullets and such. There is some difference between bullets and
ships colliding. Bullets have a damage amount and everything has a
health. When a bullet collides, it will kill its self as well as
inflicting damage on the other object. If the player and an enemy
collide they both die instantly. However this may change.